Switch the static site generation system to zola
The previous implementation was based on Hakyll and was both a bit hacky and out
of date.  Zola is more opinionated, has fewer features, and is less
customizable.  I appreciate that now.

There were only a few very minor content tweaks required.  The new layouts are a
bit different but are more visually appealing.  There is some work to do on the
theme, but it is already better than it was.

There is also some work to do to redirect the links, but that can be done at the
webserver level.